refactor: customTheme を独立化
detail ページと review ページで共通して使用するため
