+++ src/components/detail/TimelineItem.tsx
@@ -4,12 +4,7 @@ import { VFC } from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 
-import CheckboxTheme from "@/components/detail/theme/CheckboxTheme";
-import CodeTheme from "@/components/detail/theme/CodeTheme";
-import DividerTheme from "@/components/detail/theme/DividerTheme";
-import HeadingTheme from "@/components/detail/theme/HeadingTheme";
-import LinkTheme from "@/components/detail/theme/LinkTheme";
-import QuoteTheme from "@/components/detail/theme/QuoteTheme";
+import customTheme from "@/markdown/customTheme";
 
 type CustomProps = {
   comment: string | null;
@@ -18,20 +13,6 @@ type CustomProps = {
 type Props = BoxProps & CustomProps;
 
 const TimelineItem: VFC<Props> = ({ comment, ...boxProps }) => {
-  const customTheme: any = {
-    h1: HeadingTheme,
-    h2: HeadingTheme,
-    h3: HeadingTheme,
-    h4: HeadingTheme,
-    h5: HeadingTheme,
-    h6: HeadingTheme,
-    code: CodeTheme,
-    blockquote: QuoteTheme,
-    hr: DividerTheme,
-    a: LinkTheme,
-    input: CheckboxTheme,
-  };
-
   if (comment) {
     return (
       <Box {...boxProps}>
